add vacancy and admin

+++ src/by/bsu/hr/logic/DeleteUserLogic.java
@@ -6,7 +6,9 @@ import by.bsu.hr.dao.UserDAO;
 public class DeleteUserLogic {
     public static void deleteUser(int userId) throws LogicException {
         try {
-            UserDAO.deleteUser(userId);
+            if(userId!=1){
+                UserDAO.deleteUser(userId);
+            }
         } catch (DAOException e) {
             throw new LogicException("Error delete user",e);
         }
